pyproject.toml
[project]
name = "openenv-mediaops-crm-env"
version = "1.0.0"
description = "MediaOps-CRM-Env: Media agency operations RL environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
server = "my_env_name.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["my_env_name", "my_env_name.server"]
package-dir = { "my_env_name" = ".", "my_env_name.server" = "server" }